Working database check for keys and hasvisited
I implemented the check for if a player needs a key to enter the room. It calls the database for the "true" or "false" string, then checks the fake inventory currently. It also checks if the player has visited the room and displays a different description instead of the long one.